// shared-code/src/shared/enums/connection-types-enum.ts
export enum ConnectionTypesEnum {
  postgres = 'postgres',
  mysql = 'mysql',
  mysql2 = 'mysql2',
  oracledb = 'oracledb',
  mssql = 'mssql',
  ibmdb2 = 'ibmdb2',
  mongodb = 'mongodb',
  dynamodb = 'dynamodb',
  elasticsearch = 'elasticsearch',
  cassandra = 'cassandra',
  redis = 'redis',
  clickhouse = 'clickhouse',
  agent_postgres = 'agent_postgres',
  agent_mysql = 'agent_mysql',
  agent_oracledb = 'agent_oracledb',
  agent_mssql = 'agent_mssql',
  agent_ibmdb2 = 'agent_ibmdb2',
  agent_mongodb = 'agent_mongodb',
  agent_cassandra = 'agent_cassandra',
  agent_redis = 'agent_redis',
  agent_clickhouse = 'agent_clickhouse',
}

export enum ConnectionTypeTestEnum {
  postgres = 'postgres',
  mysql = 'mysql',
  mysql2 = 'mysql2',
  oracledb = 'oracledb',
  mssql = 'mssql',
  ibmdb2 = 'ibmdb2',
  mongodb = 'mongodb',
  dynamodb = 'dynamodb',
  elasticsearch = 'elasticsearch',
  cassandra = 'cassandra',
  redis = 'redis',
  clickhouse = 'clickhouse',
  agent_postgres = 'agent_postgres',
  agent_mysql = 'agent_mysql',
  agent_oracledb = 'agent_oracledb',
  agent_mssql = 'agent_mssql',
  agent_ibmdb2 = 'agent_ibmdb2',
  agent_mongodb = 'agent_mongodb',
  agent_elasticsearch = 'agent_elasticsearch',
  agent_cassandra = 'agent_cassandra',
  agent_redis = 'agent_redis',
  agent_clickhouse = 'agent_clickhouse',
}